refactor(restaurant): Create policy directory
