+++ rider/src/main/java/msa/rider/exception/RiderNonexistentException.java
@@ -1,0 +1,2 @@
+package msa.rider.exception;public class RiderNonexistentException {
+}
