+++ rider/src/main/java/msa/rider/policy/OrderStatusChangePolicy.java
@@ -1,0 +1,2 @@
+package msa.rider.policy;public class OrderStatusChangePolicy {
+}
